#23 - Consolidated array typings into 'fjl/src/types'.
