+++ packages/fjl/src/list/any.ts
@@ -1,7 +1,6 @@
 import {curry, CurryOf2} from "../function/curry";
 import {length} from "./length";
-import {PredForIndexable} from "./types";
-import {Indexable} from "../types";
+import {PredForIndexable, Indexable} from "../types";
 import {keys} from "../platform/object";
 
 export type Any<Pred, Functor> = CurryOf2<Pred, Functor, boolean>;
