+++ packages/fjl/src/list/dropWhile.ts
@@ -2,9 +2,8 @@ import {curry, CurryOf2} from "../function/curry";
 import {length} from "./length";
 import {findIndexWhere} from "./utils";
 import {slice} from "../platform/slice";
-import {PredForSlice} from "./types";
 import {sliceFrom} from "./utils/sliceFrom";
-import {Slice} from "../types";
+import {PredForSlice, Slice} from "../types";
 
 type DropWhile<T> = CurryOf2<PredForSlice<T>, Slice<T>, Slice<T>>;
