+++ packages/fjl/src/list/foldl1.ts
@@ -1,8 +1,7 @@
 import {curry, CurryOf2} from "../function/curry";
 import {uncons} from "./uncons";
 import {reduce} from "./utils/reduce";
-import {ReduceOp} from "../platform/types";
-import {Slice} from "../types";
+import {ReduceOp, Slice} from "../types";
 
 export const
