+++ packages/fjl/src/list/utils/reduce.ts
@@ -1,7 +1,6 @@
 import {$reduceUntil, reduceUntil} from "./reduceUntil";
 import {alwaysFalse} from "../../boolean/alwaysFalse";
-import {ReduceOp} from "../../platform";
-import {Indexable} from "../../types";
+import {ReduceOp, Indexable} from "../../types";
 
 export const
