+++ packages/fjl/src/platform/index.ts
@@ -5,4 +5,3 @@ export * from './function';
 export * from './object';
 export * from './slice';
 export * from './string';
-export * from './types';
